requirements.txt
APScheduler==3.6.3
attrs==22.1.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest~=7.1.2
python-telegram-bot==13.13
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests~=2.28.1
six==1.16.0
tomli==2.0.1
tornado==6.2
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11

XlsxWriter~=3.0.7